Add files via upload
A script that parses the geospatial videos and converts each individual frame into a separate JPEG. The frames are then indexed, and the model that identifies the tents in input images is imported, and the total tents detected in each frame are made into a list. The relevant data is then written into a CSV file. The resulting output is a CSV file with the index of the frames in each video and the number of tents detected in that particular frame.