Add files via upload
A script to delete the contents of the "Output Frames" folder if the production of video frames needs to be halted for any reason.
This is a cumbersome process if done manually, so this script helps save time. Frames can number in the tens of thousands if the frame generation script is interrupted.